backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
email-validator==2.1.0
tensorflow==2.15.0
Pillow==10.1.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
fastapi-mail